Allow flatten with doc comments
